fix for 64bit rfind npos issue

+++ src/learning/Classifier.cpp
@@ -85,7 +85,7 @@ void Classifier::setPredictSingleClass(bool flag) {
 
 
 std::string Classifier::getSubFilename(const std::string &baseFilename, unsigned int i) const {
-  unsigned int ind = baseFilename.rfind(".");
+  size_t ind = baseFilename.rfind(".");
   assert(ind != std::string::npos);
   return baseFilename.substr(0,ind) + "-" + boost::lexical_cast<std::string>(i) + baseFilename.substr(ind);
 }
